+++ src/main/kotlin/com/github/sleepypanda/feesh/features/overlays/ArchfiendDiceProfitTracker.kt
@@ -13,6 +13,7 @@ import com.github.sleepypanda.feesh.utils.ChatUtils
 import com.github.sleepypanda.feesh.utils.RegisterUtils
 import com.github.sleepypanda.feesh.utils.PriceUtils
 import com.github.sleepypanda.feesh.utils.gui.FeeshGui
+import com.github.sleepypanda.feesh.utils.gui.LineInfo
 import com.github.sleepypanda.feesh.utils.gui.GuiButton
 import com.github.sleepypanda.feesh.utils.enums.ColorCodes.*
 import com.github.sleepypanda.feesh.utils.enums.FormattingCodes.*
@@ -354,7 +355,7 @@ object ArchfiendDiceProfitTracker {
             lines.add("")
             lines.add("${AQUA}${BOLD}Total profit: ${profitColor}${profitShort}")
 
-            gui.setLines(lines)
+            gui.setLines(lines.map { LineInfo(it) })
             gui.setButtons(listOf(
                 GuiButton(0, "${GRAY}[Click to show $nextModeText${GRAY}]", { toggleViewMode() }),
                 GuiButton(1, "${GRAY}[${RED}Click to reset${GRAY}]", { resetArchfiendDiceProfitTracker(false, getCurrentViewMode()) })