+++ src/main/kotlin/com/github/sleepypanda/feesh/features/overlays/CrimsonIsleTracker.kt
@@ -20,6 +20,7 @@ import com.github.sleepypanda.feesh.events.models.GameClosedEvent
 import com.github.sleepypanda.feesh.events.models.OwnSeaCreatureCaughtEvent
 import com.github.sleepypanda.feesh.events.models.RareDropEvent
 import com.github.sleepypanda.feesh.utils.gui.FeeshGui
+import com.github.sleepypanda.feesh.utils.gui.LineInfo
 import com.github.sleepypanda.feesh.utils.gui.GuiButton
 import com.github.sleepypanda.feesh.settings.categories.Overlays
 import com.github.sleepypanda.feesh.utils.data.PersistentDataManager
@@ -252,7 +253,7 @@ object CrimsonIsleTracker {
         lines.addAll(data.lordJawbus.getOverlayText(lordJawbus.displayName))
         lines.addAll(data.radioactiveVials.getOverlayText(radioactiveVial.displayName, lordJawbus.displayName))
 
-        gui.setLines(lines)
+        gui.setLines(lines.map { LineInfo(it) })
         gui.setButtons(listOf(GuiButton(0, "${GRAY}[${RED}Click to reset${GRAY}]", { resetCrimsonIsleTracker(false) })))
     }