+++ src/main/kotlin/com/github/sleepypanda/feesh/features/overlays/SeaCreaturesTracker.kt
@@ -16,6 +16,7 @@ import com.github.sleepypanda.feesh.utils.RegisterUtils
 import com.github.sleepypanda.feesh.utils.gui.FeeshGui
 import com.github.sleepypanda.feesh.utils.gui.GuiButton
 import com.github.sleepypanda.feesh.utils.gui.LineAction
+import com.github.sleepypanda.feesh.utils.gui.LineInfo
 import com.github.sleepypanda.feesh.utils.enums.ColorCodes.*
 import com.github.sleepypanda.feesh.utils.enums.FormattingCodes.*
 import com.github.sleepypanda.feesh.utils.FishingHookUtils
@@ -417,7 +418,7 @@ object SeaCreaturesTracker {
                 }
             } else 0.0
 
-            TrackerLineEntry(
+            TrackerSeaCreatureEntry(
                 seaCreature = key,
                 seaCreatureInfo = seaCreatureInfo,
                 amount = value.amount,
@@ -440,11 +441,11 @@ object SeaCreaturesTracker {
             SeaCreaturesTrackerSorting.CATCHES_COUNT_DESC -> entries.sortedByDescending { it.amount }
             SeaCreaturesTrackerSorting.CATCHES_COUNT_ASC -> entries.sortedBy { it.amount }
             SeaCreaturesTrackerSorting.RARITY_ASC -> entries.sortedWith(
-                compareBy<TrackerLineEntry> { CommonUtils.getRarityNumericCode(it.seaCreatureInfo.rarityColorCode) }
+                compareBy<TrackerSeaCreatureEntry> { CommonUtils.getRarityNumericCode(it.seaCreatureInfo.rarityColorCode) }
                 .thenByDescending { it.amount }
             )
             SeaCreaturesTrackerSorting.RARITY_DESC -> entries.sortedWith(
-                compareByDescending<TrackerLineEntry> { CommonUtils.getRarityNumericCode(it.seaCreatureInfo.rarityColorCode) }
+                compareByDescending<TrackerSeaCreatureEntry> { CommonUtils.getRarityNumericCode(it.seaCreatureInfo.rarityColorCode) }
                 .thenByDescending { it.amount }
             )
         }
@@ -453,15 +454,15 @@ object SeaCreaturesTracker {
         val entriesToShow = sortedEntries.take(maxLines)
         val hiddenEntries = sortedEntries.drop(maxLines)
 
-        val lines = mutableListOf<String>()
         val nextMode = if (viewMode == ViewMode.SESSION) ViewMode.TOTAL else ViewMode.SESSION
         val nextModeText = getViewModeDisplayText(nextMode)
 
         val viewModeText = getViewModeDisplayText(viewMode)
-        lines.add("$baseTitle $viewModeText")
+        val lines = mutableListOf<LineInfo>()
+        lines.add(LineInfo("$baseTitle $viewModeText"))
 
         entriesToShow.forEach { entry ->
-            val seaCreatureText = if (entry.seaCreatureInfo.isRare) "${entry.seaCreatureInfo.boldDisplayName}" else "${entry.seaCreatureInfo.displayName}"
+            val seaCreatureText = if (entry.seaCreatureInfo.isRare) entry.seaCreatureInfo.boldDisplayName else entry.seaCreatureInfo.displayName
             val countText = "${WHITE}${entry.formattedAmount}"
             val percentText = if (showPercentage) " ${GRAY}${entry.formattedPercent}" else ""
             
@@ -469,39 +470,31 @@ object SeaCreaturesTracker {
                 " ${DARK_GRAY}| ${GRAY}DH: ${WHITE}${entry.formattedDoubleHookAmount} ${GRAY}${entry.formattedDoubleHookPercent}"
             } else ""
 
-            lines.add("${GRAY}- $seaCreatureText${GRAY}: $countText$percentText$doubleHookText")
+            lines.add(
+                LineInfo(
+                    text = "${GRAY}- $seaCreatureText${GRAY}: $countText$percentText$doubleHookText",
+                    tooltip = getSeaCreatureTooltip(entry),
+                    actions = listOf(
+                        LineAction("${GRAY}[${GREEN}+${GRAY}]") { onLineSeaCreatureIncrease(entry.seaCreature) },
+                        LineAction("${GRAY}[${RED}-${GRAY}]") { onLineSeaCreatureDecrease(entry.seaCreature) },
+                        LineAction("${GRAY}[${RED}x${GRAY}]") { onLineSeaCreatureDelete(entry.seaCreature) }
+                    )
+                )
+            )
         }
 
         if (hiddenEntries.isNotEmpty()) {
             val otherSeaCreaturesCount = CommonUtils.formatNumberWithSpaces(hiddenEntries.sumOf { it.amount })
             val otherSeaCreaturesTypes = CommonUtils.formatNumberWithSpaces(hiddenEntries.size)
-            lines.add("${GRAY}- Other sea creatures of ${WHITE}$otherSeaCreaturesTypes ${GRAY}types: ${WHITE}$otherSeaCreaturesCount")
+            lines.add(LineInfo("${GRAY}- Other sea creatures of ${WHITE}$otherSeaCreaturesTypes ${GRAY}types: ${WHITE}$otherSeaCreaturesCount"))
         }
 
         val totalLineText = getTotalLineText(displayMode, sourceObj, sortedEntries)
-        val totalLineTooltip = getTotalLineTooltip()
-        lines.add("")
-        lines.add(totalLineText)
+        val totalLineTooltip = getTotalLineTooltip(sourceObj)
+        lines.add(LineInfo(""))
+        lines.add(LineInfo(totalLineText, tooltip = totalLineTooltip))
 
         gui.setLines(lines)
-
-        val lineIndexToActions = mutableMapOf<Int, List<LineAction>>()
-        val lineIndexToTooltip = mutableMapOf<Int, Component>()
-        val buttonLinesCount = 2 // Buttons count (view mode, reset)
-        val titleLineIndex = buttonLinesCount
-        entriesToShow.forEachIndexed { index, entry ->
-            val creatureLineIndex = titleLineIndex + 1 + index
-            lineIndexToActions[creatureLineIndex] = listOf(
-                LineAction("${GRAY}[${GREEN}+${GRAY}]") { onLineSeaCreatureIncrease(entry.seaCreature) },
-                LineAction("${GRAY}[${RED}-${GRAY}]") { onLineSeaCreatureDecrease(entry.seaCreature) },
-                LineAction("${GRAY}[${RED}x${GRAY}]") { onLineSeaCreatureDelete(entry.seaCreature) }
-            )
-            lineIndexToTooltip[creatureLineIndex] = getSeaCreatureTooltip(entry)
-        }
-        
-        lineIndexToTooltip[titleLineIndex + entriesToShow.size + 1] = totalLineTooltip // Other SC line
-        gui.setTooltips(lineIndexToTooltip)
-        gui.setLineActions(lineIndexToActions)
 
         gui.setButtons(listOf(
             GuiButton(0, "${GRAY}[Click to show ${nextModeText}${GRAY}]", { toggleViewMode() }),
@@ -509,40 +502,43 @@ object SeaCreaturesTracker {
         ))
     }
 
-    private fun getTotalLineText(displayMode: SeaCreaturesTrackerDisplayMode, sourceObj: SeaCreaturesData, sortedEntries: List<TrackerLineEntry>): String {
+    private fun getSeaCreatureTooltip(entry: TrackerSeaCreatureEntry): List<Component> {
+        val dhText = if (entry.seaCreatureInfo.canBeDoubleHooked) {
+            "${GRAY}DH: ${WHITE}${entry.formattedDoubleHookAmount} ${GRAY}(${WHITE}${entry.formattedDoubleHookPercent} ${GRAY}of catches)"
+        } else ""
+        val lines = listOf(
+            entry.seaCreatureInfo.displayName,
+            "${GRAY}Total: ${WHITE}${entry.formattedAmount}",
+            dhText
+        )
+        return lines.map { Component.literal(it) }
+    }
+
+    private fun getTotalLineText(displayMode: SeaCreaturesTrackerDisplayMode, sourceObj: SeaCreaturesData, sortedEntries: List<TrackerSeaCreatureEntry>): String {
+        val allTotal = sourceObj.totalCount
+        val allTotalFormatted = CommonUtils.formatNumberWithSpaces(allTotal)
         when (displayMode) {
             SeaCreaturesTrackerDisplayMode.ALL -> {
-                val totalCount = sourceObj.totalCount
-                return "${AQUA}Total: ${WHITE}${CommonUtils.formatNumberWithSpaces(totalCount)}"
+                return "${AQUA}Total: ${WHITE}${allTotalFormatted}"
             }
             SeaCreaturesTrackerDisplayMode.ONLY_RARE -> {
-                val totalCount = sortedEntries.sumOf { it.amount }
                 val rareTotal = sortedEntries.sumOf { it.amount }
-                return "${AQUA}Total: ${WHITE}${CommonUtils.formatNumberWithSpaces(rareTotal)} ${GRAY}rare out of ${WHITE}${CommonUtils.formatNumberWithSpaces(sourceObj.totalCount)}"
+                val rareTotalFormatted = CommonUtils.formatNumberWithSpaces(rareTotal)
+                return "${AQUA}Total: ${WHITE}${rareTotalFormatted} ${GRAY}rare out of ${WHITE}${allTotalFormatted}"
             }
         }
     }
 
-    private fun getSeaCreatureTooltip(entry: TrackerLineEntry): Component {
-        return Component.literal(
-"""
-${entry.seaCreatureInfo.boldDisplayName}
-${GRAY}Caught: ${WHITE}${entry.formattedAmount} ${GRAY} sea creature(s)
-${GRAY}Double hook: ${WHITE}${entry.formattedDoubleHookAmount} ${GRAY} time(s) - ${WHITE}${entry.formattedDoubleHookPercent}%
-""".trim()
+    private fun getTotalLineTooltip(sourceObj: SeaCreaturesData): List<Component> {
+        val allTotal = sourceObj.totalCount
+        val allTotalFormatted = CommonUtils.formatNumberWithSpaces(allTotal)
+        val lines = listOf(
+            "${GRAY}Total: ${WHITE}${allTotalFormatted}",
         )
+        return lines.map { Component.literal(it) }
     }
 
-    private fun getTotalLineTooltip(): Component {
-        return Component.literal(
-"""
-${AQUA}Total: ???
-${AQUA}Rare: ???
-""".trim()
-        )
-    }
-
-    private data class TrackerLineEntry(
+    private data class TrackerSeaCreatureEntry(
         val seaCreature: String,
         val seaCreatureInfo: SeaCreatures.SeaCreatureInfo,
         val amount: Int,