+++ docs/CHANGELOG.md
@@ -14,6 +14,8 @@ Released on: ???
   - Basically now you can toggle those functionalities for any sea creature.
   - Only "rare" creatures are selected by default.
 - Added alert/pchat message when you cocoon a specific sea creature.
+- Added tooltips with various extra info when you hover over a line in Sea creature tracker (while in Inventory).
+  - So even if you hide % or DH info in the tracker, you still can see it on hover.
 - Added personal best tracking for double hook streak.
 - Added alert after personal SC cap is reached (based on chat message).